requirements.txt
streamlit
keras
tensorflow
numpy
opencv-python
pillow
